+++ CHANGELOG.md
@@ -166,7 +166,7 @@
 - Wait for Devdocs Assets to be loaded before generating the devdocs site !3662 @SaiCode
 - Updated Clean script with clean node_modules !3688 @SaiCode
 - Removed some unnecessary database queries #1995 !3709 !3718 !3725 @alex.simm
-- Added a command to move old uploaded pictures to the API #1182 #1440 !3631 @alex.simm
+- Added a command to move old uploaded pictures to the API #1182 #1440 !3631 !3822 @alex.simm
 - Fix missing sql dump creation in test script !3739 @chriswalg
 - Added dataSource.xml and sqldialects.xml for JetBrains IDEs !3755 @chriswalg
 - Removed unused hidden dialog !3761 @alex.simm